Prepare StopSearch for search by line
* Move all hooks and files into the same base directory.
* Rename files to allow for "alternative" implementation of the search.
* Split code to smaller units.
