+++ ui/src/components/stop-registry/search/StopSearchResultsPage.tsx
@@ -1,31 +1,20 @@
 import { useTranslation } from 'react-i18next';
-import {
-  usePagination,
-  useStopSearch,
-  useStopSearchResults,
-} from '../../../hooks';
-import { Container, Row, Visible } from '../../../layoutComponents';
-import { CloseIconButton, Pagination } from '../../../uiComponents';
-import { LoadingWrapper } from '../../../uiComponents/LoadingWrapper';
-import { sortAlphabetically } from '../../../utils';
+import { Container, Row } from '../../../layoutComponents';
+import { CloseIconButton } from '../../../uiComponents';
+import { StopSearchByStopResults } from './by-stop';
 import { StopSearchBar } from './StopSearchBar';
-import { StopSearchResultList } from './StopSearchResultList';
+import { useStopSearch } from './useStopSearch';
+
+const testIds = {
+  container: 'StopSearchResultsPage::Container',
+  closeButton: 'StopSearchResultsPage::closeButton',
+};
 
 export const StopSearchResultPage = (): React.ReactElement => {
-  const { handleClose } = useStopSearch();
   const { t } = useTranslation();
-  const { stops, loading, resultCount } = useStopSearchResults();
 
-  const itemsPerPage = 10;
-  const { getPaginatedData } = usePagination();
-  const sortedStopsByLabel = sortAlphabetically(stops, 'label');
-  const displayedStops = getPaginatedData(sortedStopsByLabel, itemsPerPage);
-
-  const testIds = {
-    container: 'StopSearchResultsPage::Container',
-    closeButton: 'StopSearchResultsPage::closeButton',
-    loadingSearchResults: 'LoadingWrapper::loadingStopSearchResults',
-  };
+  const { handleClose, searchConditions } = useStopSearch();
+  console.log({ searchConditions });
 
   return (
     <Container testId={testIds.container}>
@@ -41,25 +30,7 @@ export const StopSearchResultPage = (): React.ReactElement => {
         />
       </Row>
       <StopSearchBar />
-      <LoadingWrapper
-        className="flex justify-center"
-        loadingText={t('search.searching')}
-        loading={loading}
-        testId={testIds.loadingSearchResults}
-      >
-        {/* TODO: Search result filter input */}
-        {/* TODO: Selection toolbar */}
-        <StopSearchResultList stops={displayedStops} />
-        <Visible visible={!!resultCount}>
-          <div className="grid grid-cols-4">
-            <Pagination
-              className="col-span-2 col-start-2 pt-4"
-              itemsPerPage={itemsPerPage}
-              totalItemsCount={resultCount}
-            />
-          </div>
-        </Visible>
-      </LoadingWrapper>
+      <StopSearchByStopResults />
     </Container>
   );
 };
