+++ ui/src/components/stop-registry/search/types/StopPlaceSearchRowDetails.ts
@@ -1,0 +1,5 @@
+export type StopPlaceSearchRowDetails = {
+  netexId?: string | null;
+  nameFin?: string | null;
+  nameSwe?: string | null;
+};
