+++ ui/src/components/stop-registry/search/types/StopSearchRow.ts
@@ -1,0 +1,6 @@
+import { StopTableRowFragment } from '../../../../generated/graphql';
+import { StopPlaceSearchRowDetails } from './StopPlaceSearchRowDetails';
+
+export type StopSearchRow = StopTableRowFragment & {
+  stop_place: StopPlaceSearchRowDetails;
+};
