+++ ui/src/components/stop-registry/search/types/index.ts
@@ -1,0 +1,2 @@
+export * from './StopPlaceSearchRowDetails';
+export * from './StopSearchRow';
