+++ ui/src/components/stop-registry/stop-areas/stop-area-details/components/StopAreaMemberStopRows.tsx
@@ -4,9 +4,8 @@ import { FC } from 'react';
 import { useTranslation } from 'react-i18next';
 import { twMerge } from 'tailwind-merge';
 import { StopAreaDetailsFragment } from '../../../../../generated/graphql';
-import { StopSearchRow } from '../../../../../hooks';
 import { StopAreaFormMember } from '../../../../forms/stop-area';
-import { StopTableRow } from '../../../search';
+import { StopSearchRow, StopTableRow } from '../../../search';
 import { LocatorActionButton } from '../../../search/StopTableRow/ActionButtons/LocatorActionButton';
 import { OpenDetailsPage } from '../../../search/StopTableRow/MenuItems/OpenDetailsPage';
 import { ShowOnMap } from '../../../search/StopTableRow/MenuItems/ShowOnMap';
