--- ui/src/hooks/stop-registry/search/index.ts
@@ -1,3 +1,0 @@
-export * from './useStopSearch';
-export * from './useStopSearchQueryParser';
-export * from './useStopSearchResults';
